--- src/database/database.providers.ts
@@ -1,21 +1,0 @@
-import { Sequelize } from 'sequelize-typescript';
-import { User } from 'src/users/user.model';
-
-export const databaseProviders = [
-  {
-    provide: 'SEQUELIZE',
-    useFactory: async () => {
-      const sequelize = new Sequelize({
-        dialect: 'postgres',
-        host: 'localhost',
-        port: 3306,
-        username: 'root',
-        password: 'password',
-        database: 'nest',
-      });
-      sequelize.addModels([User]);
-      await sequelize.sync();
-      return sequelize;
-    },
-  },
-];
