+++ src/main.ts
@@ -11,6 +11,7 @@ async function bootstrap() {
   .setVersion('1.0')
   .addTag('Sistema')
   .build();
+  
   const document = SwaggerModule.createDocument(app, config);
   SwaggerModule.setup('api', app, document);
