+++ src/modules/users/contracts/dto/create.ts
@@ -1,0 +1,5 @@
+export class CreateUserDto {
+  nome: string;
+  password: string;
+  cpf: string;
+}
