--- src/users/create-user.dto.ts
@@ -1,5 +1,0 @@
-export class CreateUserDto {
-    nome: string;
-    password: string;
-    cpf: string;
-  }
