--- src/users/user.provider.ts
@@ -1,8 +1,0 @@
-import { User } from "./user.model";
-
-export const usersProviders = [
-  {
-    provide: 'USERS_REPOSITORY',
-    useValue: User,
-  },
-];
